backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
claude-sdk==0.7.0
anthropic==0.7.0
pymupdf==1.23.8
pdfminer.six==20221105
python-docx==0.8.11
supabase==1.3.4
httpx==0.25.2
aiofiles==23.2.1
aiohttp==3.9.1
cors==1.0.1
python-cors==0.0.2
fitz==0.0.1.dev2
PyPDF2==3.0.1
